refactor: useModalContext 를 hook으로 분리 및 import 경로 수정
