+++ components/src/lib/components/common/Modal.Container.tsx
@@ -1,7 +1,7 @@
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
 
-import { useModalContext } from './useModalContext';
+import { useModalContext } from '../../hooks/useModalContext';
 
 type ModalContainerProps = {
   /**
